+++ Test/CoreSDK.Test/Net46/Extensibility/Implementation/RichPayloadEventSourceTest.cs
@@ -2,11 +2,18 @@
 {
     using System;
     using System.Collections.Generic;
+
+#if NET40
+    using Microsoft.Diagnostics.Tracing;
+#else
     using System.Diagnostics.Tracing;
+#endif
     using System.Linq;
     using Channel;
     using Microsoft.ApplicationInsights.DataContracts;
+
     using Microsoft.VisualStudio.TestTools.UnitTesting;
+    using System.Reflection;
 
     /// <summary>
     /// Tests the rich payload event source tracking.
@@ -23,6 +30,7 @@
             this.DoTracking(
                 RichPayloadEventSource.Keywords.Requests,
                 new RequestTelemetry("TestRequest", DateTimeOffset.Now, TimeSpan.FromMilliseconds(10), "200", true),
+                typeof(External.RequestData),
                 (client, item) => { client.TrackRequest((RequestTelemetry)item); });
         }
 
@@ -35,6 +43,7 @@
             this.DoTracking(
                 RichPayloadEventSource.Keywords.Traces,
                 new TraceTelemetry("TestTrace", SeverityLevel.Information),
+                typeof(External.MessageData),
                 (client, item) => { client.TrackTrace((TraceTelemetry)item); });
         }
 
@@ -47,6 +56,7 @@
             this.DoTracking(
                 RichPayloadEventSource.Keywords.Events,
                 new EventTelemetry("TestEvent"),
+                typeof(External.EventData),
                 (client, item) => { client.TrackEvent((EventTelemetry)item); });
         }
 
@@ -56,9 +66,13 @@
         [TestMethod]
         public void RichPayloadEventSourceExceptionSentTest()
         {
+            var exceptionTelemetry = new ExceptionTelemetry(new SystemException("Test"));
+            exceptionTelemetry.Data.exceptions[0].parsedStack = new External.StackFrame[] { new External.StackFrame() };
+
             this.DoTracking(
                 RichPayloadEventSource.Keywords.Exceptions,
-                new ExceptionTelemetry(new SystemException("Test")),
+                exceptionTelemetry,
+                typeof(External.ExceptionData),
                 (client, item) => { client.TrackException((ExceptionTelemetry)item); });
         }
 
@@ -71,6 +85,7 @@
             this.DoTracking(
                 RichPayloadEventSource.Keywords.Metrics,
                 new MetricTelemetry("TestMetric", 1),
+                typeof(External.MetricData),
                 (client, item) => { client.TrackMetric((MetricTelemetry)item); });
         }
 
@@ -83,6 +98,7 @@
             this.DoTracking(
                 RichPayloadEventSource.Keywords.Dependencies,
                 new DependencyTelemetry("Custom", "Target", "TestDependency", "TestCommand", DateTimeOffset.Now, TimeSpan.Zero, "200", true),
+                typeof(External.RemoteDependencyData),
                 (client, item) => { client.TrackDependency((DependencyTelemetry)item); });
         }
 
@@ -95,7 +111,38 @@
             this.DoTracking(
                 RichPayloadEventSource.Keywords.PageViews,
                 new PageViewTelemetry("TestPage"),
+                typeof(External.PageViewData),
                 (client, item) => { client.TrackPageView((PageViewTelemetry)item); });
+        }
+
+        /// <summary>
+        /// Tests tracking session state telemetry.
+        /// </summary>
+        [TestMethod]
+        public void RichPayloadEventSourceSessionStateSentTest()
+        {
+#pragma warning disable 618
+            this.DoTracking(
+                RichPayloadEventSource.Keywords.Events,
+                new SessionStateTelemetry(new SessionState()),
+                typeof(External.EventData),
+                (client, item) => { client.Track((SessionStateTelemetry)item); });
+#pragma warning restore 618
+        }
+
+            /// <summary>
+            /// Tests tracking session state telemetry.
+            /// </summary>
+        [TestMethod]
+        public void RichPayloadEventSourceSessionPerformanceCounterTest()
+        {
+#pragma warning disable 618
+            this.DoTracking(
+                RichPayloadEventSource.Keywords.Metrics,
+                new PerformanceCounterTelemetry("TestCategory", "TestCounter", "TestInstance", 1.0),
+                typeof(External.MetricData),
+                (client, item) => { client.Track((PerformanceCounterTelemetry)item); });
+#pragma warning restore 618
         }
 
         /// <summary>
@@ -104,52 +151,133 @@
         /// <param name="keywords">The event keywords to enable.</param>
         /// <param name="item">The telemetry item to track.</param>
         /// <param name="track">The tracking callback to execute.</param>
-        private void DoTracking(EventKeywords keywords, ITelemetry item, Action<TelemetryClient, ITelemetry> track)
+        private void DoTracking(EventKeywords keywords, ITelemetry item, Type dataType, Action<TelemetryClient, ITelemetry> track)
         {
-            var client = new TelemetryClient();
-            client.InstrumentationKey = Guid.NewGuid().ToString();
-
-            using (var listener = new Microsoft.ApplicationInsights.TestFramework.TestEventListener())
+            if (IsRunningOnEnvironmentSupportingRichPayloadEventSource())
             {
-                listener.EnableEvents(RichPayloadEventSource.Log.EventSourceInternal, EventLevel.Verbose, keywords);
+                var client = new TelemetryClient();
+                client.InstrumentationKey = Guid.NewGuid().ToString();
 
-                item.Context.Properties.Add("property1", "value1");
-                item.Context.User.Id = "testUserId";
-                item.Context.Operation.Id = Guid.NewGuid().ToString();
-
-                track(client, item);
-
-                var actualEvent = listener.Messages.FirstOrDefault();
-
-                Assert.IsNotNull(actualEvent);
-                Assert.AreEqual(client.InstrumentationKey, actualEvent.Payload[0]);
-
-                int keysFound = 0;
-                object[] tags = actualEvent.Payload[1] as object[];
-                foreach (object tagObject in tags)
+                using (var listener = new Microsoft.ApplicationInsights.TestFramework.TestEventListener())
                 {
-                    Dictionary<string, object> tag = (Dictionary<string, object>)tagObject;
-                    Assert.IsNotNull(tag);
-                    string key = (string)tag["Key"];
-                    object value = tag["Value"];
-                    if (!string.IsNullOrWhiteSpace(key))
+                    listener.EnableEvents(RichPayloadEventSource.Log.EventSourceInternal, EventLevel.Verbose, keywords);
+
+                    item.Context.Properties.Add("property1", "value1");
+                    item.Context.User.Id = "testUserId";
+                    item.Context.Operation.Id = Guid.NewGuid().ToString();
+
+                    track(client, item);
+
+                    var actualEvent = listener.Messages.FirstOrDefault();
+
+                    Assert.IsNotNull(actualEvent);
+                    Assert.AreEqual(client.InstrumentationKey, actualEvent.Payload[0]);
+
+                    int keysFound = 0;
+                    object[] tags = actualEvent.Payload[1] as object[];
+                    foreach (object tagObject in tags)
                     {
-                        if (key == "ai.user.id")
+                        Dictionary<string, object> tag = (Dictionary<string, object>)tagObject;
+                        Assert.IsNotNull(tag);
+                        string key = (string)tag["Key"];
+                        object value = tag["Value"];
+                        if (!string.IsNullOrWhiteSpace(key))
                         {
-                            Assert.AreEqual("testUserId", value);
-                            ++keysFound;
+                            if (key == "ai.user.id")
+                            {
+                                Assert.AreEqual("testUserId", value);
+                                ++keysFound;
+                            }
+                            else if (key == "ai.operation.id")
+                            {
+                                Assert.AreEqual(item.Context.Operation.Id, value);
+                                ++keysFound;
+                            }
                         }
-                        else if (key == "ai.operation.id")
+                    }
+
+                    Assert.AreEqual(2, keysFound);
+                    Assert.IsNotNull(actualEvent.Payload[2]);
+
+                    var expectedProperties = dataType.GetProperties().AsEnumerable();
+                    var actualPropertiesPayload = (IDictionary<string, object>)actualEvent.Payload[2];
+                    VerifyEventPayload(expectedProperties, actualPropertiesPayload);
+                }
+            }
+            else
+            {
+                // 4.5 doesn't have RichPayload events
+                Assert.IsNull(RichPayloadEventSource.Log.EventSourceInternal);
+            }
+        }
+
+        private static void VerifyEventPayload(IEnumerable<PropertyInfo> expectedProperties, IDictionary<string, object> actualEventPayload)
+        {
+            var actualProperties = actualEventPayload.Keys.Select(k => new { Key = k, Value = actualEventPayload[k] });
+
+            Assert.IsTrue(expectedProperties.Count() == actualProperties.Count());
+            var expectedPropertiesEnumerator = expectedProperties.GetEnumerator();
+            var actualPropertiesEnumerator = actualProperties.GetEnumerator();
+            while (expectedPropertiesEnumerator.MoveNext() && actualPropertiesEnumerator.MoveNext())
+            {
+                var expectedProperty = expectedPropertiesEnumerator.Current;
+                var actualProperty = actualPropertiesEnumerator.Current;
+                Assert.AreEqual(expectedPropertiesEnumerator.Current.Name, actualPropertiesEnumerator.Current.Key);
+
+                if (!expectedProperty.PropertyType.IsValueType
+                    && !expectedProperty.PropertyType.IsPrimitive
+                    && expectedProperty.PropertyType != typeof(string))
+                {
+                    if (expectedProperty.PropertyType.IsClass)
+                    {
+                        VerifyEventPayload(expectedProperty.PropertyType.GetProperties().AsEnumerable(), (IDictionary<string, object>)actualProperty.Value);
+                    }
+                    else
+                    {
+                        var enumerableType = GetEnumerableType(expectedProperty.PropertyType);
+                        Assert.IsNotNull(enumerableType);
+                        Assert.IsTrue(actualProperty.Value.GetType().IsArray);
+
+                        if (enumerableType.IsClass)
                         {
-                            Assert.AreEqual(item.Context.Operation.Id, value);
-                            ++keysFound;
+                            VerifyEventPayload(enumerableType.GetProperties().AsEnumerable(), (IDictionary<string, object>)((object[])actualProperty.Value)[0]);
                         }
                     }
                 }
 
-                Assert.AreEqual(2, keysFound);
-                Assert.IsNotNull(actualEvent.Payload[2]);
             }
+        }
+
+        private static Type GetEnumerableType(Type type)
+        {
+            foreach (Type intType in type.GetInterfaces())
+            {
+                if (intType.IsGenericType && intType.GetGenericTypeDefinition() == typeof(IEnumerable<>))
+                {
+                    return intType.GetGenericArguments()[0];
+                }
+            }
+            return null;
+        }
+
+        private static bool IsRunningOnEnvironmentSupportingRichPayloadEventSource()
+        {
+#if NET40
+            // NET40 version uses EventSource in Microsoft.Diagnostics which supports RichPayloadEvent on .NET Framework 4.0/4.5/4.6+
+            return true;
+#else
+            // Other versions depend on EventSource in .Net Framework 4.6+
+            string productVersionString = System.Diagnostics.FileVersionInfo.GetVersionInfo(typeof(object).Assembly.Location).ProductVersion;
+
+            Version ver;
+            if (!Version.TryParse(productVersionString, out ver))
+            {
+                Assert.Fail("Unable to determine .net framework version");
+            }
+
+            var ver46 = new Version(4, 6, 0, 0);
+            return ver >= ver46;
+#endif
         }
     }
 }