+++ src/Core/Managed/Shared/TelemetryClient.cs
@@ -352,7 +352,7 @@
                 // invokes the Process in the first processor in the chain
                 this.configuration.TelemetryProcessorChain.Process(telemetry);
 
-#if NET46
+#if !CORE_PCL
                 // logs rich payload ETW event for any partners to process it
                 RichPayloadEventSource.Log.Process(telemetry);
 #endif